Change the implementation of the short-circuit.
Using "return true" inside of an if else tree does not work.
Changed the implementation so that after the if else tree for the
hit_chance, there is another if to check if the hit_chance is 0.
May not be worth it, but this is how it will be done for now.