Collision functions can now be specified in blocks.
This frees up the specification of collision funciton to the
subclasses of Body_3D.  A side effect of this change is that
Body_3D can not be used directly,as it will define no collision
functions.  This is acceptable as it was never going to be used
directly in this program anyway.  One of it's subclasses should
always be used instead.