Rename ChipmunkInterfaceMod.rb to Space_3D.rb
The only thing left in this file once all the other stuff was moved out was the 
Space_3D code, thus, the name change.  Now all of the Chipmunk specific code is in
one directory rather than in one file.  It was never really confined in one file anyway,
so this way is better.