Require set as CP::Space_3D now uses sets instead of arrays.
This is actually unnecessary, as the imports should really all be called in init_game.
This file can not be executed by itself, so there really is no reason to do it this way.
However, I prefer this style so I can track the dependencies of different files.
Perhaps this should be changed to a comment later.