Add methods to get the width and the height of the Entity.
These values are currently based on the dimentions of the sprite.  Thus, they are 
in pixles.  This is strange, as the other units are in meters.  This means that the units must be specified as an argument of these methods.  However, this makes it
so that to_meters and to_px are not used universally.  If the physics engine is overhauled, this problem with be resolved.