Compute the CP::Space_3D timestep based on the time between frames.
Before, the timestep was a fixed 1/60th of a second.  This meant that if the 
framerate dropped, the game would slow down.  This will not work in a 
multiplayer game with real-time combat.  The frames need to be dropped 
and the current position of things displayed.  Thus, the new behavior.