Restructure Ground.rb to hold all the parts needed to form the ground.
The three classes, Ground::Renderer, Ground::Texture, and 
Ground::Texture_Map combine to form a cohesive high-level unit that can be
drawn as if it were one infinitely scalable Gosu::Image.  Like wallpaper, or
a bucket fill, it could be theoretically applied to a space of any given 
dimensions.  The number of textures should be low, but the result should 
be as seemling random as possible.