Change Shadow to be one class, which takes one argument.
Shadow now takes one argument, the Entity from which to make a shadow of.
From this is it can figure out the class of the Entity, which it uses as a key in
the hash table of all Gosu::Image objects used for Shadow.  It can also find
the location of where the shadow should be drawn.