Change Ellipse and Blip to use units of meters instead of pixels.
This change was made to allow the usage of the new #draw method.  This
required some reworking of code.  In the process, the variable storing the
center of the ellipse was removed, and now the position of the player
is used instead.  Before they were the same value, being copied over every
frame.