Change conditionals to include when @@time is equal to the expected value.
Forgot that the value would be an integer, and not a double or float, and thus
this might actually be possible.  It would be more likely if the Timer was 
updated constantly in a separate thread.  As I have considered making this
engine multi-threaded to take advantage of modern hardware, I want to make
the transition as easy as possible if I choose that route in the future.