Change subsprites argument to a standard Array.
Before it had been *subsprites in the argument list, which meant that if
simply an array was supplied as an argument, then subsprites would become
an Array with an Array inside of it.  This is not the intended behavior.  As this
class is never to be used directly, the argument was thus changed.