Refine Ubuntu installer, getting dependencies in a single apt-get or gem call.
The installer now uses bash to concat the strings of the dependencies,
and then install them in one call to apt-get or gem instead of having
multiple calls.  Thus, the user will be able to see how much memory is
required up front, and then say yes or no accordingly.