Golem Smash complete
There was also some paint weight work done on the front end of the
shoulders where the pec muscles connect to the shoulders.  Still a quad
or so on either side that gets smushed at full arm raising height, but I
don't think it can be helped at such an extreme angle/pose.